Hotfix Commit #1 : Fixing an issue where the application would crash immediately when trying to enter an expense with a date in the future.
Issue: The issue lied in the way the review methos were structured, in the previous state, the application performed a check for the Date only after having created the expense, which triggered an IllegalArgumentException to be thrown from the internal Expense.java class.

Solution: Reorganized validation scheme in the setUpButtonsForInteraction() method within the AddExpenseView.java file such that validation reviews the Date before attempting to create the Expense.